- 205 тестов ОК. 3 фейла...

+++ ewm-main-service/src/main/java/ru/practicum/mnsvc/dto/participation/EventRequestStatusUpdateDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class EventRequestStatusUpdateDto {
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private ParticipationState status;
 
 }
